tests/wpt/tests/css/css-scroll-snap/scroll-initial-target/parsing/scroll-initial-target-valid.html

<!DOCTYPE html>
<html>

<head>
  <meta charset="utf-8">
  <title> CSS Scroll Snap 2 Test: scroll-initial-target-* with valid values</title>
  <link rel="help" href="https://drafts.csswg.org/css-scroll-snap-2/#scroll-initial-target">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/parsing-testcommon.js"></script>
</head>

<body>
  <script>
    test_valid_value("scroll-initial-target", "none");
    test_valid_value("scroll-initial-target", "nearest");
  </script>
</body>

</html>
